+++ mcfp/__init__.py
@@ -1,1 +1,7 @@
+from .compiler import Compiler
+
 __version__ = "0.1.0"
+
+__all__ = [
+    'Compiler',
+]
