+++ pyproject.toml
@@ -2,17 +2,18 @@
 name = "mcfp"
 version = "0.1.0"
 description = ""
-authors = [
-    {name = "yuyi2439",email = "yuyi2439@qq.com"}
-]
-license = {text = "MIT"}
+authors = [{ name = "yuyi2439", email = "yuyi2439@qq.com" }]
+license = { text = "MIT" }
 readme = "README.md"
 requires-python = ">=3.11"
-dependencies = [
-    "pytest (>=8.3.5,<9.0.0)"
-]
 
+[tool.poetry.group.dev.dependencies]
+pytest = "^8.3.5"
 
 [build-system]
 requires = ["poetry-core>=2.0.0,<3.0.0"]
 build-backend = "poetry.core.masonry.api"
+
+
+[tool.pytest.ini_options]
+pythonpath = "."
